Move `advance_slot` to its own file.
